pyproject.toml
[project]
name = "automation-v1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "msoffcrypto-tool>=5.4.2",
    "openai>=2.9.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pyinstaller>=6.17.0",
    "streamlit>=1.51.0",
]
